Fix application
TLDR; this fixes everything.

Unfortunately, updating Rails 3.2.1 to 3.2.11 blindly breaks show routes.  I'm guessing this is partially because the Journey router updated, among other significant changes in the 3.2.x release.

Also, there seems to have been some breaking changes in `activerecord-postgres-hstore`, so I locked it to what was in your checked-in Gemfile.  Including the latest version of this gem will break `schema.rb`.